Determine fitness based on the generation
Return a fitness value that includes either the max X value or the max Y
value, depending on the number of generation that we are at.

Also removed the function for drawing the cluster center phenotypes of
the last generation, since we need to implement it using the
'Population' instance.