Get partially-correct algebraic expression
Convert positions to a partially-correct 'sympy' expression. The reason
that it is not completely correct is that we cannot express X as
'x = ...' when 'm' == 0 in 'y = mx + n'. Hence, I assumed that when the
line is parallel to X axis, the line is 'y = x + n'. However, this
misses that if we assume this, Y must not be the same after the end of
that line segment.